+++ src/com/intellij/idea/plugin/hybris/tools/remote/console/actions/handler/HybrisConsoleExecuteActionHandler.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -121,7 +121,6 @@ class HybrisConsoleExecuteActionHandler(private val project: Project,
         // Process input and add to history
         val document = console.currentEditor.document
         val textForHistory = document.text
-        console.preProcessors().forEach { processor -> console.setInputText(processor.process(console)) }
 
         val query = document.text
         val range = TextRange(0, document.textLength)